+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - issue-15: Removed last references to on-prem l2ss and podaac-tools
 ### Fixed
+- issue-33: Fixed collection resolution displaying multiple resolutions and scaling
 
 
 ## [4.16.0]